pyproject.toml
[tool.poetry]
name = "poetry-python-best-practices"
version = "0.1.0"
description = "Template for python best practices such as mypy, coverage, flake8 integration and pytests including mock tests."
authors = ["G Shiva"]
license = "Apache"
readme = "README.md"
packages = [{include = "poetry_python_best_practices"}]

[tool.poetry.dependencies]
python = "^3.11"
argparse = "^1.4.0"
gitpython = "^3.1.31"
click = "^8.1.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-mock = "*"
black = "^23.1.0"
mypy = "^1.1.1"
coverage = "^7.2.1"
interrogate = "^1.5.0"
sphinx = "^6.1.3"
sphinx-rtd-theme = "^1.2.0"


[[tool.poetry.source]]
name = "mypypi"
url = "https://pypi.org/simple/"
default = false
secondary = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true

